+++ Makefile
@@ -10,7 +10,7 @@
 # #*******************************************************************************
 
 
-SOURCES=$(wildcard *.py)
+SOURCES:=$(shell find storage_api -name "*.py")
 
 clean:
 	find . -name \*.pyc -o -name \*.pyo -o -name __pycache__ -exec rm -rf {} +
